# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Open API documentation for project Album-singers",
        "contact": {
            "email": "galina@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Laravel Swagger API server"
        }
    ],
    "paths": {
...
        "/singers": {
            "get": {
                "tags": [
                    "Singers"
                ],
                "summary": "Display a listing of the resource",
                "operationId": "singersAll",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Singers"
                ],
                "summary": "Create yet another singer record",
                "operationId": "singerCreate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Singer name",
                                        "type": "string",
                                        "example": "Alex"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Created siccessful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SingerShowResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/singers/{id}": {
            "get": {
                "tags": [
                    "Singers"
                ],
                "summary": "Display a listing of the resource",
                "operationId": "singersOne",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of singer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "5"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SingerShowResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Singers"
                ],
                "summary": "Delete singer by ID",
                "operationId": "singersDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of singer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successful"
                    },
                    "404": {
                        "description": "Model not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Singers"
                ],
                "summary": "Update yet another singer record",
                "operationId": "singerUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of singer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "5"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Singer name",
                                        "type": "string",
                                        "example": "Alex"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Created siccessful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SingerShowResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Model not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "SingerShowResource": {
                "title": "Singer model",
                "description": "Some simple request createa as example",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "5"
                    },
                    "name": {
                        "description": "The name",
                        "type": "string",
                        "example": "Alex"
                    }
                },
                "type": "object"
            }
        }
    },
...
}